+++ scripts/run_full_opencl_sim_test.sh
@@ -14,10 +14,13 @@ $SCRIPT_DIR_PATH/setup_bsp.sh
 
 cd $ROOT_PROJECT_PATH/example_designs/mem_bandwidth
 rm -fr bin/mem_bandwidth
-aoc device/mem_bandwidth.cl --board skx_fpga_dcp_ddr -o bin/mem_bandwidth.aocx
-#aoc device/mem_bandwidth.cl --board skx_fpga_dcp_svm -o bin/mem_bandwidth.aocx
-rm -fr mem_bandwidth_comp
-mv bin/mem_bandwidth mem_bandwidth_comp
+if [ ! -f bin/mem_bandwidth.aocx ]; then
+	echo "Running AOC..."
+	aoc device/mem_bandwidth.cl --board skx_fpga_dcp_ddr -o bin/mem_bandwidth.aocx
+	#aoc device/mem_bandwidth.cl --board skx_fpga_dcp_svm -o bin/mem_bandwidth.aocx
+	rm -fr mem_bandwidth_comp
+	mv bin/mem_bandwidth mem_bandwidth_comp
+fi
 aocl program acl0 bin/mem_bandwidth.aocx
 make
 ./bin/mem_bandwidth 1
